exitをreturnにしたことによって処理が進んでしまう箇所のエラーハンドリング

+++ 04.class/actions/deleteMemo.js
@@ -1,12 +1,19 @@
 import { selectMemo } from "../helpers/memoHelpers.js";
 
 export async function deleteMemo(database) {
-  const selectedMemoId = await selectMemo(
-    database,
-    "削除するメモを選んでください:",
-  );
+  let selectedMemoId;
+  try {
+    selectedMemoId = await selectMemo(
+      database,
+      "削除するメモを選んでください:",
+    );
+  } catch (err) {
+    return console.log(err.message);
+  }
 
   const memo = await database.fetchMemoById(selectedMemoId);
+  if (!memo) return;
+
   await database.deleteMemoById(selectedMemoId);
   process.stdout.write(`${memo.content}\nメモを削除しました。\n`);
 }
